// src/web/lgse/src/app/models/api/resources/get-active-mprn-response.ts
export class GetActiveMPRnList {
    id: string;
    status: number;
    incidentId: string;
    cell: string;
    zone: string;
    zoneManagerName:string;
    cellManagerName:string;
    priorityCustomer: boolean;
    postcode: string;
    country: string;
    dependentLocality: string;
    localityName: string;
    postTown: string;
    dependentStreet: string
    principalStreet: string;
    buildingNumber: string;
    subBuildingName: string;
    buildingName: string;
    mprn: string;
    checked: boolean;
    assignedResourceCount: number;
    isIsolated: boolean;
    latestStatus:string;
    latestStatusId:string;
    latestSubStatus:string;
    latestSubStatusId:string;
    notes:string;
    notesCount:number;

}